Update PlayerScript.cs
In the update(), looking for GetKey(KeyCode.LeftShift())
When this happens, I'm changing the transform.localScale of the Player.  releasing returns scale to original parameters.
Next phase will be to try and isolate this function to the left/right thrusters of the ship.
After this, link this action to an increase in speed to simulate active thrusters.