Initial commit for Shield Strength branch
Added logic to the Damage() method to detect hits to the shield.  When it's detected, the alpha of the sprite renderer is changed then updated.
There are three hits before shields go out.
Capturing a new Shield PowerUp resets the hits on the shield and bring shield back to full strength.