Initial Commit - Lateral Laser
New laser weapon for Player: laser shot that fires laterally out the port/starbord sides of the ship, in conjunction with normal or triple shot lasers.
Created a new Prefab for the lateral laser.
Added logic in the Laser script for when lateral shot is active.
Added logic in PlayerScript to instantiate lateral laser shots.

Left to do: create power-up, and maybe modify the Prefab to make the shot more daunting...