Completed a wave system
Modified the Game Manager to keep track of waves.
Each wave can be customised to have a different enemy, rates of fire, spawn time, diff qty of enemies, diff speeds.
Also added more aural warnings and adjusted code to accommodate new waves.
adjusted the UI to show max qty of ammo and missiles.
Started experimenting with Raycast2D with a new empty game object as a child to the Player with a basic script attached.