Raming Enemy Test
Finally managed to create an enemy prefab variant with a raycast script as a child.  The raycast detects Playe, then increases the speed of the Enemy ship for as long as the Player is lit up by the raycast.  When Player moves out of the way, enemy speed returns to assigned speed for enemy ships in the wave as per Game Manager data.