Added company intro screen
Added a company label intro screen (which is what I call my bootstrapper screen).
Basically a scene with text/logo which fades/zooms in with the music, and continues to zoom in/fade out before switching to the Main Menu scene.
Controlled the fade/zoom using a loop and "Lerping" the values of the text alpha and the main camera field of view over time.

Also did a few minor tweaks in the game, and fixed a couple of bugs in order to reset default starting game values when Game Over and restarting without going through the Main Menu scene.